Put each serialiser in its own module and made serialisers into a package
